Create README.md
Actually this is a reverse shell.What is a reverse shell: Shell shoveling, in network security, refers to the act of redirecting the input and output of a shell to a service so that it can be remotely accessed. In computing, the most basic method of interfacing with the operating system is the shell.I have created this system for one of my fiverr client.This contain two apps 1/A admin app to control victim device 2/A reverse Shell (victim app) to fire up the services and host the data to firebase database.
What the device services this system can execute
1/Access the victim camera and take snap without his awareness
2/listen through microphone
3/Read sms and send sms
4/Read Contact and meke phone call
5/run in background as a spy and it will also remove it's icon so it will never appear in screen

The admin can control all the things mentioned above from his dashboard with some button click.But because of security reason I have uploaded version 1.0 which only can Read and Send Sms.